Fix - Status enum type 을 GameStatus 객체 내부로 이동하여 캡슐화
