--- src/main/java/User.java
@@ -1,3 +1,0 @@
-public class User {
-
-}
